Move test data downloader and data typo generator to a shared subcrate so I can reuse it for benching
