+++ Cargo.toml
@@ -23,7 +23,7 @@ criterion = "0.2"
 tempfile = "3.0.2"
 rand = "0.5.0"
 lazy_static = "1.0"
-libflate = "0.1.11"
+test_utils = { path = "test_utils" }
 
 [features]
 default = ["mmap"]
